reorganised rangevoting into planning app, begun tabs, added wrapper-related code (incomplete)
