+++ apps/range-voting/app/.babelrc
@@ -1,0 +1,28 @@
+{
+  "presets": [
+    [
+      "env",
+      {
+        "modules": false,
+        "targets": {
+          "browsers": [
+            "> 1%",
+            "last 3 versions",
+            "ie >= 9",
+            "ios >= 8",
+            "android >= 4.2"
+          ]
+        },
+        "useBuiltIns": false
+      }
+    ],
+    "stage-2"
+  ],
+  "plugins": [
+    ["styled-components", {
+      "displayName": true
+    }],
+    "transform-runtime"
+  ]
+}
+
