+++ apps/range-voting/app/package.json
@@ -4,23 +4,53 @@
   "private": true,
   "description": "",
   "dependencies": {
-    "@aragon/ui": "^0.4.0",
-    "@aragon/client": "^1.0.0-beta.2",
+    "@aragon/ui": "^0.11.0",
+    "@aragon/client": "^1.0.0-beta.7",
     "styled-components": "^3.1.6",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-motion": "^0.5.2"
+    "react-motion": "^0.5.2",
+    "date-fns": "^2.0.0-alpha.7",
+    "prop-types": "^15.6.0",
+    "react-blockies": "^1.2.2",
+    "seed-random": "^2.2.0"
   },
   "devDependencies": {
-    "webpack": "^3.8.1",
-    "react-scripts": "1.0.17"
+    "@aragon/react-scripts": "^1.1.0-aragon.4",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.3",
+    "babel-plugin-styled-components": "^1.5.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "eslint": "4.15.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-config-standard-react": "^5.0.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-prettier": "^2.3.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.5.1",
+    "eslint-plugin-standard": "^3.0.1",
+    "parcel-bundler": "^1.7.0",
+    "prettier": "^1.8.2"
   },
   "scripts": {
+    "lint": "eslint ./src",
     "sync-assets": "rsync -rtu \"$(dirname $(node -p 'require.resolve(\"@aragon/ui\")'))/\" public/aragon-ui",
     "start": "npm run sync-assets && PORT=3001 react-scripts start",
     "build": "npm run sync-assets && react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "license": "GPL-3.0"
+  "license": "GPL-3.0",
+  "proxy": "http://localhost:3001",
+  "browserslist": [
+    ">2%",
+    "last 1 edge versions",
+    "not ie > 0",
+    "not op_mini all"
+  ]
 }
