+++ apps/range-voting/app/src/range-voting/templates.js
