+++ apps/range-voting/app/src/range-voting/templates/issuecuration/index.js
