+++ apps/range-voting/app/src/range-voting/templates/payoutengine/ConfigureVotingConditions.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Field, DropDown } from '@aragon/ui'
-import { lerp } from '../../math-utils'
-import { noop } from '../../utils'
-import { Main, Content, Title, Subtitle, Hint } from '../../style'
+import { Main, Content, Title, Subtitle, Hint } from '../../../style'
 import { votePermissionItems, voteWeightItems, voteOutcomeItems } from './voting-conditions'
+import { lerp } from '../../../utils/math-utils'
+import { noop } from '../../../utils/utils'
 
 class ConfigureVotingConditions extends React.Component {
   static defaultProps = {
