+++ apps/range-voting/app/src/range-voting/templates/payoutengine/ConfigureVotingDefaults.js
@@ -1,8 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Field, TextInput, Text, theme } from '@aragon/ui'
-import { lerp } from '../../math-utils'
-import { noop } from '../../utils'
+import { Field, TextInput } from '@aragon/ui'
+import { Main, Content, Title, Subtitle, Hint } from '../../../style'
+import { lerp } from '../../../utils/math-utils'
+import { noop } from '../../../utils/utils'
 
 class ConfigureVotingDefaults extends React.Component {
   static defaultProps = {
@@ -75,17 +76,17 @@ class ConfigureVotingDefaultsContent extends React.PureComponent {
     } = this.props
     return (
       <Content>
-        <Title>Democracy Project</Title>
+        <Title>Payout Engine</Title>
         <StepContainer>
           <SubmitForm onSubmit={onSubmit} innerRef={formRef}>
-            <p style={{ textAlign: 'center' }}>
-              <Text size="large" color={theme.textSecondary} align="center">
-                Choose your voting settings below. You can’t change these later,
-                so pick carefully.
-              </Text>
-            </p>
+            <Subtitle>
+              Choose your voting settings below. You can’t change these later, so pick carefully.
+            </Subtitle>
             <Fields>
               <Fields.PercentageField label="Support">
+                <Hint>
+                  Percentage of votes required to validate an option
+                </Hint>
                 <SymbolInput
                   placeholder="e.g. 50"
                   value={fields.support === -1 ? '' : fields.support}
@@ -93,6 +94,9 @@ class ConfigureVotingDefaultsContent extends React.PureComponent {
                 />
               </Fields.PercentageField>
               <Fields.PercentageField label="Min. Quorum">
+                <Hint>
+                  Percentage of token supply required to participate in proposals
+                </Hint>
                 <SymbolInput
                   placeholder="e.g. 15"
                   value={fields.minQuorum === -1 ? '' : fields.minQuorum}
@@ -100,6 +104,9 @@ class ConfigureVotingDefaultsContent extends React.PureComponent {
                 />
               </Fields.PercentageField>
               <Fields.HoursField label="Vote Duration">
+                <Hint>
+                  Duration each vote is active for
+                </Hint>
                 <SymbolInput
                   placeholder="e.g. 24"
                   onChange={handleVoteDurationChange}
@@ -121,29 +128,6 @@ const SubmitForm = ({ children, innerRef = noop, ...props }) => (
   </form>
 )
 
-const Main = styled.div`
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  padding: 100px;
-  padding-top: 140px;
-`
-
-const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`
-
-const Title = styled.h1`
-  text-align: center;
-  font-size: 37px;
-  margin-bottom: 100px;
-`
-
 const StepContainer = styled.div`
   display: flex;
   align-items: flex-start;
@@ -159,15 +143,11 @@ const SymbolInput = styled(TextInput)`
 `
 
 const Fields = styled.div`
-  display: flex;
   justify-content: center;
   margin-top: 40px;
 `
 Fields.Field = styled(Field)`
   position: relative;
-  & + & {
-    margin-left: 55px;
-  }
   &:after {
     position: absolute;
     bottom: 6px;
