+++ apps/range-voting/app/src/range-voting/templates/payoutengine/index.js
