+++ apps/range-voting/app/src/range-voting/templates/payoutengine/voting-conditions.js
