+++ apps/range-voting/app/src/range-voting/vote-settings.js
