+++ apps/range-voting/app/src/range-voting/vote-types.js
@@ -1,0 +1,5 @@
+export const VOTE_ABSENT = Symbol('VOTE_ABSENT')
+export const VOTE_STATUS_ONGOING = Symbol('VOTE_STATUS_ONGOING')
+export const VOTE_STATUS_REJECTED = Symbol('VOTE_STATUS_REJECTED')
+export const VOTE_STATUS_ACCEPTED = Symbol('VOTE_STATUS_ACCEPTED')
+
