+++ apps/range-voting/app/src/screens/Dashboard.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import styled from 'styled-components'
+import { EmptyStateCard } from '@aragon/ui'
+import emptyIcon from '../assets/empty-card-icon.svg'
+
+const EmptyIcon = () => <img src={emptyIcon} alt="" />
+
+const Content = ({ onActivate }) => (
+  <Main>
+    <EmptyStateCard
+      icon={EmptyIcon}
+      title="Dashboard"
+      text="placeholder for real dashboard"
+      actionText="New Project"
+      onActivate={onActivate}
+    />
+  </Main>
+)
+
+const Main = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-grow: 1;
+`
+
+class Dashboard extends React.Component {
+  render () {
+    const { name, active } = this.props
+    return (<Content />)
+  }
+}
+
+export default Dashboard
