+++ apps/range-voting/app/src/screens/Tools.js
@@ -1,0 +1,27 @@
+import React from 'react'
+import styled from 'styled-components'
+import { EmptyStateCard } from '@aragon/ui'
+import emptyIcon from '../assets/empty-card-icon.svg'
+
+const EmptyIcon = () => <img src={emptyIcon} alt="" />
+
+const Tools = ({ onActivate }) => (
+  <Main>
+    <EmptyStateCard
+      icon={EmptyIcon}
+      title="Tools"
+      text="placeholder for tools"
+      actionText="New Project"
+      onActivate={onActivate}
+    />
+  </Main>
+)
+
+const Main = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-grow: 1;
+`
+
+export default Tools
