+++ apps/range-voting/app/src/script.js
@@ -1,7 +1,7 @@
 import Aragon from '@aragon/client'
 import { combineLatest } from './rxjs'
-import voteSettings, { hasLoadedVoteSettings } from './vote-settings'
-import { EMPTY_CALLSCRIPT } from './vote-utils'
+import voteSettings, { hasLoadedVoteSettings } from './range-voting/vote-settings'
+import { EMPTY_CALLSCRIPT } from './range-voting/vote-utils'
 
 const app = new Aragon()
