+++ apps/range-voting/app/src/utils/utils.js
