+++ apps/range-voting/package.json
@@ -1,19 +1,28 @@
 {
   "name": "@tpt/apps-range-voting",
   "version": "0.0.1",
+  "main": "index.js",
   "description": "",
   "dependencies": {
     "@aragon/os": "^3.1.2"
   },
   "devDependencies": {
-    "@aragon/cli": "^2.0.0",
+    "@aragon/cli": "^2.0.4",
     "@aragon/test-helpers": "^1.0.0",
+    "babel-helpers": "^6.24.1",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-preset-stage-3": "^6.17.0",
+    "babel-register": "^6.26.0",
+    "eth-gas-reporter": "^0.1.1",
+    "ethereumjs-testrpc": "^6.0.2",
     "ganache-cli": "^6.0.3",
-    "truffle-hdwallet-provider": "0.0.3",
     "solidity-coverage": "0.3.5",
-    "solium": "^0.5.5",
+    "solidity-sha3": "^0.4.1",
+    "solium": "^1.0.4",
     "truffle": "4.0.5",
-    "ethereumjs-abi": "^0.6.4"
+    "truffle-hdwallet-provider": "0.0.3"
   },
   "scripts": {
     "test": "TRUFFLE_TEST=true npm run ganache-cli:dev",
